Remove `ignore()` method from subscriber api
This method had some potential blowups in that since it overwrote
the previous version of the ignore method, a subscriber could
only be resgistered to one hub at a time. This may or may not be
a desirable thing, but I don't think it should be the place of
this library to determine that.

Another thing is that this helps us keep the API relatively simple
(one method). It's now much easier to explain what a subscriber is:
an object that implements a receive method.