Add initial versions of subscribe() and emit()
I'm thinking of adding an array argument to each of these methods.
subscribe() definitely needs it, emit() might be better off without it.
Emissions should do one thing so that they trigger one type of event.
If we need to trigger several at a time, then we can do so with
multiple calls to emit().