Rename Hub api methods
This is just from a practical standpoint, but the shorter names are easier
to work with. The switch from `emit()` to `dispatch()` is out of the need
to separate this pub/sub pattern from an observer pattern using something
like EventEmitter. When they are in the same code base, it can get
confusing...