Update subscribers to be functions
This simplifies a lot of the code and means I don't have to worry about
checking for a `receive()` method. I can just call the subscriber
directly. This means we can get rid of the sub module, since that was its
only purpose.

I also removed the `matches_pattern` helper and replaced it with a
simpler to read wildcard module.